Test item count passed
