+++ src/modules/itemCounter.js
@@ -1,0 +1,8 @@
+const movies = (arr) => {
+    let length;
+    arr.forEach((element, i) => {
+      length = i + 1;
+    });
+    return length;
+  };
+  export default movies;
